tests/proton_ci/requirements.txt
PyGithub>=1.58.0
unidiff>=0.7.0
boto3==1.20.4